+++ experimental/dds/tree2/src/test/feature-libraries/editable-tree/editableTree.editing.spec.ts
@@ -24,7 +24,6 @@ import {
 	treeStatus,
 	TreeStatus,
 	FieldSchema,
-	on,
 } from "../../../feature-libraries";
 import { viewWithContent } from "../../utils";
 import {
@@ -301,141 +300,6 @@ describe("editable-tree: editing", () => {
 				"2": 3,
 			});
 		}
-	});
-
-	it.only("beforeChange/afterChange events fire the expected number of times", () => {
-		const tree = viewWithContent({ schema: fullSchemaData, initialTree: getPerson() });
-		const person = tree.root as EditableTree;
-
-		let beforeChangePersonCount = 0;
-		let afterChangePersonCount = 0;
-		let beforeChangeAddressCount = 0;
-		let afterChangeAddressCount = 0;
-
-		person[on]("beforeChange", (event) => {
-			beforeChangePersonCount++;
-		});
-		person[on]("afterChange", (event) => {
-			afterChangePersonCount++;
-		});
-
-		assert.strictEqual(beforeChangePersonCount, 0);
-		assert.strictEqual(afterChangePersonCount, 0);
-
-		// Update age; should fire events on the person node.
-		person.age = brand<Int32>(32);
-
-		assert.strictEqual(beforeChangePersonCount, 1);
-		assert.strictEqual(afterChangePersonCount, 1);
-
-		// Update address; should fire events on the person node.
-		// This also lets us put listeners on it, otherwise get complaints that person.address might be undefined below.
-		person.address = {
-			zip: "99999",
-			street: "foo",
-			phones: [12345],
-		} as unknown as Address; // TODO: fix up these strong types to reflect unwrapping
-
-		assert.strictEqual(beforeChangePersonCount, 2);
-		assert.strictEqual(afterChangePersonCount, 2);
-
-		person.address[on]("beforeChange", (event) => {
-			beforeChangeAddressCount++;
-		});
-		person.address[on]("afterChange", (event) => {
-			afterChangeAddressCount++;
-		});
-
-		assert.strictEqual(beforeChangeAddressCount, 0);
-		assert.strictEqual(afterChangeAddressCount, 0);
-
-		// Replace zip in address; should fire events on the address node and the person node.
-		person.address.zip = brand<Int32>(12345);
-
-		assert.strictEqual(beforeChangePersonCount, 3);
-		assert.strictEqual(afterChangePersonCount, 3);
-		assert.strictEqual(beforeChangeAddressCount, 1);
-		assert.strictEqual(afterChangeAddressCount, 1);
-
-		// Replace the whole address; should fire events on the person node.
-		person.address = {
-			zip: "99999",
-			street: "foo",
-			phones: [12345],
-		} as unknown as Address; // TODO: fix up these strong types to reflect unwrapping
-
-		assert.strictEqual(beforeChangePersonCount, 4);
-		assert.strictEqual(afterChangePersonCount, 4);
-		// No events should have fired on the old address node.
-		assert.strictEqual(beforeChangeAddressCount, 1);
-		assert.strictEqual(afterChangeAddressCount, 1);
-
-		// Replace zip in new address node; should fire events on the person node (but not on the old address node)
-		person.address.zip = brand<Int32>(23456);
-
-		assert.strictEqual(beforeChangePersonCount, 5);
-		assert.strictEqual(afterChangePersonCount, 5);
-		assert.strictEqual(beforeChangeAddressCount, 1);
-		assert.strictEqual(afterChangeAddressCount, 1);
-	});
-
-	it.only("beforeChange/afterChange fire in the expected order and always together", () => {
-		const tree = viewWithContent({ schema: fullSchemaData, initialTree: getPerson() });
-		const person = tree.root as EditableTree;
-
-		let beforeCounter = 0;
-		let afterCounter = 0;
-
-		person[on]("beforeChange", (event) => {
-			beforeCounter++;
-			assert.strictEqual(afterCounter, beforeCounter - 1, "beforeChange fired out of order");
-		});
-		person[on]("afterChange", (event) => {
-			afterCounter++;
-			assert.strictEqual(afterCounter, beforeCounter, "afterChange fired out of order");
-		});
-
-		// Make updates to the tree
-		person.age = brand<Int32>(32);
-		person.address = {
-			zip: "99999",
-			street: "foo",
-			phones: [12345],
-		} as unknown as Address; // TODO: fix up these strong types to reflect unwrapping
-		person.address.zip = brand<Int32>(12345);
-		person.address = {
-			zip: "99999",
-			street: "foo",
-			phones: [12345],
-		} as unknown as Address; // TODO: fix up these strong types to reflect unwrapping
-		person.address.zip = brand<Int32>(23456);
-
-		// Check the number of events fired is correct (otherwise the assertions in the listeners might not have ran)
-		assert.strictEqual(beforeCounter, 5);
-		assert.strictEqual(afterCounter, 5);
-	});
-
-	it.skip("leaf nodes don't emit beforeChange/afterChange when they are replaced", () => {
-		const tree = viewWithContent({ schema: fullSchemaData, initialTree: getPerson() });
-		const person = tree.root as EditableTree;
-		person.age = brand<Int32>(32); // Explicitly update age so we can attach listeners to it.
-		let beforeCounter = 0;
-		let afterCounter = 0;
-		// QUESTION
-		// Are we already not allowing leaf nodes to have listeners?
-		// `person.age[on]` doesn't work (error: "Element implicitly has an 'any' type because expression of type 'unique
-		// symbol' can't be used to index type 'number | EditableTree'")
-		// And with the cast to EditableTree: TypeError: person.age[feature_libraries_1.on] is not a function
-		(person.age as EditableTree)[on]("beforeChange", (event) => {
-			beforeCounter++;
-		});
-		(person.age as EditableTree)[on]("afterChange", (event) => {
-			afterCounter++;
-		});
-		person.age = brand<Int32>(33);
-		// Events shouldn't have fired on the original age node
-		assert.strictEqual(beforeCounter, 0);
-		assert.strictEqual(afterCounter, 0);
 	});
 
 	describe(`can move nodes`, () => {
